wip: move some files into other packages to clean up
