+++ app/src/main/java/com/chesire/malime/view/MainActivity.kt
@@ -14,11 +14,12 @@ import android.view.MenuItem
 import com.chesire.malime.R
 import com.chesire.malime.core.flags.ItemType
 import com.chesire.malime.mal.room.MalimeDatabase
-import com.chesire.malime.util.PeriodicUpdateHelper
+import com.chesire.malime.util.updateservice.PeriodicUpdateHelper
 import com.chesire.malime.util.SharedPref
 import com.chesire.malime.view.login.LoginActivity
 import com.chesire.malime.view.maldisplay.MalDisplayFragment
 import com.chesire.malime.view.preferences.PrefActivity
+import com.chesire.malime.view.preferences.SortOption
 import com.chesire.malime.view.search.SearchFragment
 import timber.log.Timber
