+++ app/src/main/java/com/chesire/malime/view/maldisplay/MalDisplayViewAdapter.kt
@@ -17,7 +17,7 @@ import com.chesire.malime.util.GlideApp
 import com.chesire.malime.util.SharedPref
 import com.chesire.malime.util.preferenceFilter
 import com.chesire.malime.util.preferenceSort
-import com.chesire.malime.view.SortOption
+import com.chesire.malime.view.preferences.SortOption
 import kotlinx.android.synthetic.main.item_malmodel.view.item_malmodel_content_layout
 import kotlinx.android.synthetic.main.item_malmodel.view.item_malmodel_image
 import kotlinx.android.synthetic.main.item_malmodel.view.item_malmodel_loading_layout
